backend/requirements.txt
flask
flask-cors
llama-index-core
llama-index-llms-ollama
llama-index-embeddings-ollama
